move log rotation to own file
